# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type TileType = typeof TileType[keyof typeof TileType];


export const TileType = {
  app: 'app',
  truenas: 'truenas',
  media: 'media',
  sonarr: 'sonarr',
} as const;

/**
 * @nullable
 */
export type TileImageFit = typeof TileImageFit[keyof typeof TileImageFit] | null;


export const TileImageFit = {
  cover: 'cover',
  contain: 'contain',
  center: 'center',
  'top-left': 'top-left',
} as const;

export interface Tile {
  id: number;
  userId: number;
  type: TileType;
  gridX: number;
  gridY: number;
  gridW: number;
  gridH: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  bgColor?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageFit?: TileImageFit;
  createdAt?: string;
}

export type TileInputType = typeof TileInputType[keyof typeof TileInputType];


export const TileInputType = {
  app: 'app',
  truenas: 'truenas',
  media: 'media',
  sonarr: 'sonarr',
} as const;

export interface TileInput {
  type: TileInputType;
  gridX: number;
  gridY: number;
  gridW: number;
  gridH: number;
  name?: string;
  url?: string;
  bgColor?: string;
  imageUrl?: string;
  imageFit?: string;
}

export interface TileUpdate {
  gridX?: number;
  gridY?: number;
  gridW?: number;
  gridH?: number;
  name?: string;
  url?: string;
  bgColor?: string;
  imageUrl?: string;
  imageFit?: string;
}

export interface LayoutItem {
  id: number;
  gridX: number;
  gridY: number;
  gridW: number;
  gridH: number;
}

export interface LayoutUpdate {
  tiles: LayoutItem[];
}

export interface UploadResult {
  url: string;
}

export interface ZfsPool {
  name: string;
  status: string;
  usedBytes: number;
  totalBytes: number;
}

export interface TruenasMetrics {
  cpuPercent: number;
  memUsedGb: number;
  memTotalGb: number;
  pools: ZfsPool[];
}

export interface MediaItem {
  id: string;
  title: string;
  type: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  thumb?: string | null;
  /** @nullable */
  addedAt?: string | null;
}

export interface SonarrQueueItem {
  id: number;
  title: string;
  status: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  size?: number | null;
}

export interface SonarrUpcomingItem {
  id: number;
  title: string;
  seriesTitle?: string;
  airDate: string;
  /** @nullable */
  seasonNumber?: number | null;
  /** @nullable */
  episodeNumber?: number | null;
}

export interface SonarrData {
  queue: SonarrQueueItem[];
  upcoming: SonarrUpcomingItem[];
}