# lib/api-zod/src/generated/types/layoutItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface LayoutItem {
  id: number;
  gridX: number;
  gridY: number;
  gridW: number;
  gridH: number;
}