# lib/api-zod/src/generated/types/layoutUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { LayoutItem } from './layoutItem';

export interface LayoutUpdate {
  tiles: LayoutItem[];
}