# lib/api-zod/src/generated/types/sonarrData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { SonarrQueueItem } from './sonarrQueueItem';
import type { SonarrUpcomingItem } from './sonarrUpcomingItem';

export interface SonarrData {
  queue: SonarrQueueItem[];
  upcoming: SonarrUpcomingItem[];
}