# lib/api-zod/src/generated/types/sonarrQueueItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SonarrQueueItem {
  id: number;
  title: string;
  status: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  size?: number | null;
}