# lib/api-zod/src/generated/types/sonarrUpcomingItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SonarrUpcomingItem {
  id: number;
  title: string;
  seriesTitle?: string;
  airDate: string;
  /** @nullable */
  seasonNumber?: number | null;
  /** @nullable */
  episodeNumber?: number | null;
}