# lib/api-zod/src/generated/types/tile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TileImageFit } from './tileImageFit';
import type { TileType } from './tileType';

export interface Tile {
  id: number;
  userId: number;
  type: TileType;
  gridX: number;
  gridY: number;
  gridW: number;
  gridH: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  bgColor?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageFit?: TileImageFit;
  createdAt?: string;
}