# lib/api-zod/src/generated/types/tileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TileUpdate {
  gridX?: number;
  gridY?: number;
  gridW?: number;
  gridH?: number;
  name?: string;
  url?: string;
  bgColor?: string;
  imageUrl?: string;
  imageFit?: string;
}