# lib/api-zod/src/generated/types/zfsPool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Homelab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ZfsPool {
  name: string;
  status: string;
  usedBytes: number;
  totalBytes: number;
}